[RTL][NTDLL] Implement NT6 locale functions and test
Implement RTL locale functions. This includes a locale table, which for now is hardcoded. On Windows Vista and later (as well as Wine) it is part of locale.nls, but we don't have support for that yet.

-Proposed changes:

    [NDK] Minor header improvements
    [NDK] Add some Vista+ Rtl locale functions
    [NTDLL_APITEST] Add test for Rtl locale functions
    [RTL] Implement locale functions/tables
    [NTDLL] Export RTL locale functions

-Tests:

These functions are NT6, so we currently do not have a way to test them either on Vista or on ReactOS. I tested them locally and all tests pass on Windows 10 and ReactOS NT6.